code cleanup for #388
