+++ src/main/java/net/sf/jabref/ProxyAuthenticator.java
@@ -1,0 +1,39 @@
+/*  Copyright (C) 2003-2015 JabRef contributors.
+    This program is free software; you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation; either version 2 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along
+    with this program; if not, write to the Free Software Foundation, Inc.,
+    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+*/
+package net.sf.jabref;
+
+import java.net.Authenticator;
+import java.net.PasswordAuthentication;
+
+public class ProxyAuthenticator extends Authenticator {
+
+    @Override
+    protected PasswordAuthentication getPasswordAuthentication() {
+        if (getRequestorType() == RequestorType.PROXY) {
+            String prot = getRequestingProtocol().toLowerCase();
+            String host = System.getProperty(prot + ".proxyHost", "");
+            String port = System.getProperty(prot + ".proxyPort", "80");
+            String user = System.getProperty(prot + ".proxyUser", "");
+            String password = System.getProperty(prot + ".proxyPassword", "");
+            if (getRequestingHost().equalsIgnoreCase(host)) {
+                if (Integer.parseInt(port) == getRequestingPort()) {
+                    return new PasswordAuthentication(user, password.toCharArray());
+                }
+            }
+        }
+        return null;
+    }
+}
